+++ src/ExecutionContext/constants.js
@@ -3,8 +3,7 @@
  * @type {Object<String>}
  */
 exports.ExecutionContextErrors = {
-    CONTEXT_ALREADY_DECLARED: 'Execution context is already declared for the given domain, use the domain option to create a separate context.',
-    CONTEXT_DOES_NOT_EXISTS: 'Execution context does not exists, please ensure to call create/run before.',
+    CONTEXT_DOES_NOT_EXIST: 'Execution context does not exists, please ensure to call create/run before.',
     MONITOR_MISS_CONFIGURATION: 'Monitoring option is off by default, please call `configure` with the proper options.'
 };
 
@@ -15,9 +14,3 @@ exports.ExecutionContextErrors = {
 exports.DEFAULT_CONFIG = {
     monitor: false
 };
-
-/**
- * The default domain to create execution context roots under.
- * @type {String}
- */
-exports.ROOT_DOMAIN = 'ROOT';