+++ src/managers/asyncHooks/constants.js
@@ -1,0 +1,7 @@
+/**
+ * The default configuration to use.
+ * @type {ExecutionContextConfig}
+ */
+exports.DEFAULT_CONFIG = {
+    monitor: false
+};